Refactoring webdeployments_configuration to its own package
