+++ genesyscloud/webdeployments_configuration/resource_genesyscloud_webdeployments_configuration_utils.go
@@ -1,0 +1,487 @@
+package webdeployments_configuration
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"strconv"
+	gcloud "terraform-provider-genesyscloud/genesyscloud"
+	lists "terraform-provider-genesyscloud/genesyscloud/util/lists"
+	"time"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/mypurecloud/platform-client-sdk-go/v105/platformclientv2"
+)
+
+func customizeConfigurationDiff(ctx context.Context, diff *schema.ResourceDiff, meta interface{}) error {
+	if len(diff.GetChangedKeysPrefix("")) > 0 {
+		// When any change is made to the configuration we automatically publish a new version, so mark the version as updated
+		// so dependent deployments will update appropriately to reference the newest version
+		diff.SetNewComputed("version")
+	}
+	return nil
+}
+
+func validateConfigurationStatusChange(k, old, new string, d *schema.ResourceData) bool {
+	// Configs start in a pending status and may not transition to active or error before we retrieve the state, so allow
+	// the status to change from pending to something less ephemeral
+	return old == "Pending"
+}
+
+func determineLatestVersion(ctx context.Context, api *platformclientv2.WebDeploymentsApi, configurationId string) string {
+	version := ""
+	_ = gcloud.WithRetries(ctx, 30*time.Second, func() *resource.RetryError {
+		versions, resp, getErr := api.GetWebdeploymentsConfigurationVersions(configurationId)
+		if getErr != nil {
+			if gcloud.IsStatus404(resp) {
+				return resource.RetryableError(fmt.Errorf("Failed to determine latest version %s", getErr))
+			}
+			log.Printf("Failed to determine latest version. Defaulting to DRAFT. Details: %s", getErr)
+			version = "draft"
+			return resource.NonRetryableError(fmt.Errorf("Failed to determine latest version %s", getErr))
+		}
+
+		maxVersion := 0
+		for _, v := range *versions.Entities {
+			if *v.Version == "DRAFT" {
+				continue
+			}
+			APIVersion, err := strconv.Atoi(*v.Version)
+			if err != nil {
+				log.Printf("Failed to convert version %s to an integer", *v.Version)
+			} else {
+				if APIVersion > maxVersion {
+					maxVersion = APIVersion
+				}
+			}
+		}
+
+		if maxVersion == 0 {
+			version = "draft"
+		}
+
+		version = strconv.Itoa(maxVersion)
+
+		return nil
+	})
+
+	return version
+}
+
+func flattenMessengerSettings(messengerSettings *platformclientv2.Messengersettings) []interface{} {
+	return []interface{}{map[string]interface{}{
+		"enabled":         messengerSettings.Enabled,
+		"styles":          flattenStyles(messengerSettings.Styles),
+		"launcher_button": flattenLauncherButton(messengerSettings.LauncherButton),
+		"file_upload":     flattenFileUpload(messengerSettings.FileUpload),
+	}}
+}
+
+func flattenCobrowseSettings(cobrowseSettings *platformclientv2.Cobrowsesettings) []interface{} {
+	return []interface{}{map[string]interface{}{
+		"enabled":             cobrowseSettings.Enabled,
+		"allow_agent_control": cobrowseSettings.AllowAgentControl,
+		"channels":            cobrowseSettings.Channels,
+		"mask_selectors":      cobrowseSettings.MaskSelectors,
+		"readonly_selectors":  cobrowseSettings.ReadonlySelectors,
+	}}
+}
+
+func flattenStyles(styles *platformclientv2.Messengerstyles) []interface{} {
+	if styles == nil {
+		return nil
+	}
+
+	return []interface{}{map[string]interface{}{
+		"primary_color": styles.PrimaryColor,
+	}}
+}
+
+func flattenLauncherButton(settings *platformclientv2.Launcherbuttonsettings) []interface{} {
+	if settings == nil {
+		return nil
+	}
+
+	return []interface{}{map[string]interface{}{
+		"visibility": settings.Visibility,
+	}}
+}
+
+func flattenFileUpload(settings *platformclientv2.Fileuploadsettings) []interface{} {
+	if settings == nil || settings.Modes == nil || len(*settings.Modes) < 1 {
+		return nil
+	}
+
+	modes := make([]map[string]interface{}, len(*settings.Modes))
+	for i, mode := range *settings.Modes {
+		modes[i] = map[string]interface{}{
+			"file_types":       *mode.FileTypes,
+			"max_file_size_kb": *mode.MaxFileSizeKB,
+		}
+	}
+
+	return []interface{}{map[string]interface{}{
+		"mode": modes,
+	}}
+}
+
+func flattenJourneyEvents(journeyEvents *platformclientv2.Journeyeventssettings) []interface{} {
+	return []interface{}{map[string]interface{}{
+		"enabled":                   journeyEvents.Enabled,
+		"excluded_query_parameters": journeyEvents.ExcludedQueryParameters,
+		"should_keep_url_fragment":  journeyEvents.ShouldKeepUrlFragment,
+		"search_query_parameters":   journeyEvents.SearchQueryParameters,
+		"pageview_config":           journeyEvents.PageviewConfig,
+		"click_event":               flattenSelectorEventTriggers(journeyEvents.ClickEvents),
+		"form_track_event":          flattenFormsTrackTriggers(journeyEvents.FormsTrackEvents),
+		"idle_event":                flattenIdleEventTriggers(journeyEvents.IdleEvents),
+		"in_viewport_event":         flattenSelectorEventTriggers(journeyEvents.InViewportEvents),
+		"scroll_depth_event":        flattenScrollPercentageEventTriggers(journeyEvents.ScrollDepthEvents),
+	}}
+}
+
+func flattenSelectorEventTriggers(triggers *[]platformclientv2.Selectoreventtrigger) []interface{} {
+	if triggers == nil || len(*triggers) < 1 {
+		return nil
+	}
+
+	result := make([]interface{}, len(*triggers))
+	for i, trigger := range *triggers {
+		result[i] = map[string]interface{}{
+			"selector":   trigger.Selector,
+			"event_name": trigger.EventName,
+		}
+	}
+	return result
+}
+
+func flattenFormsTrackTriggers(triggers *[]platformclientv2.Formstracktrigger) []interface{} {
+	if triggers == nil || len(*triggers) < 1 {
+		return nil
+	}
+
+	result := make([]interface{}, len(*triggers))
+	for i, trigger := range *triggers {
+		result[i] = map[string]interface{}{
+			"selector":                     trigger.Selector,
+			"form_name":                    trigger.FormName,
+			"capture_data_on_form_abandon": trigger.CaptureDataOnFormAbandon,
+			"capture_data_on_form_submit":  trigger.CaptureDataOnFormSubmit,
+		}
+	}
+	return result
+}
+
+func flattenIdleEventTriggers(triggers *[]platformclientv2.Idleeventtrigger) []interface{} {
+	if triggers == nil || len(*triggers) < 1 {
+		return nil
+	}
+
+	result := make([]interface{}, len(*triggers))
+	for i, trigger := range *triggers {
+		result[i] = map[string]interface{}{
+			"event_name":         trigger.EventName,
+			"idle_after_seconds": trigger.IdleAfterSeconds,
+		}
+	}
+	return result
+}
+
+func flattenScrollPercentageEventTriggers(triggers *[]platformclientv2.Scrollpercentageeventtrigger) []interface{} {
+	if triggers == nil || len(*triggers) < 1 {
+		return nil
+	}
+
+	result := make([]interface{}, len(*triggers))
+	for i, trigger := range *triggers {
+		result[i] = map[string]interface{}{
+			"event_name": trigger.EventName,
+			"percentage": trigger.Percentage,
+		}
+	}
+	return result
+}
+
+func waitForConfigurationDraftToBeActive(ctx context.Context, api *platformclientv2.WebDeploymentsApi, id string) diag.Diagnostics {
+	return gcloud.WithRetries(ctx, 30*time.Second, func() *resource.RetryError {
+		configuration, resp, err := api.GetWebdeploymentsConfigurationVersionsDraft(id)
+		if err != nil {
+			if gcloud.IsStatus404(resp) {
+				return resource.RetryableError(fmt.Errorf("Error verifying active status for new web deployment configuration %s: %s", id, err))
+			}
+			return resource.NonRetryableError(fmt.Errorf("Error verifying active status for new web deployment configuration %s: %s", id, err))
+		}
+
+		if *configuration.Status == "Active" {
+			return nil
+		}
+
+		return resource.RetryableError(fmt.Errorf("Web deployment configuration %s not active yet. Status: %s", id, *configuration.Status))
+	})
+}
+
+func readWebDeploymentConfigurationFromResourceData(d *schema.ResourceData) (string, *platformclientv2.Webdeploymentconfigurationversion) {
+	name := d.Get("name").(string)
+	languages := lists.InterfaceListToStrings(d.Get("languages").([]interface{}))
+	defaultLanguage := d.Get("default_language").(string)
+
+	inputCfg := &platformclientv2.Webdeploymentconfigurationversion{
+		Name:            &name,
+		Languages:       &languages,
+		DefaultLanguage: &defaultLanguage,
+	}
+
+	description, ok := d.Get("description").(string)
+	if ok {
+		inputCfg.Description = &description
+	}
+
+	messengerSettings := readMessengerSettings(d)
+	if messengerSettings != nil {
+		inputCfg.Messenger = messengerSettings
+	}
+
+	cobrowseSettings := readCobrowseSettings(d)
+	if cobrowseSettings != nil {
+		inputCfg.Cobrowse = cobrowseSettings
+	}
+
+	journeySettings := readJourneySettings(d)
+	if journeySettings != nil {
+		inputCfg.JourneyEvents = journeySettings
+	}
+
+	return name, inputCfg
+}
+
+func readJourneySettings(d *schema.ResourceData) *platformclientv2.Journeyeventssettings {
+	value, ok := d.GetOk("journey_events")
+	if !ok {
+		return nil
+	}
+
+	cfgs := value.([]interface{})
+	if len(cfgs) < 1 {
+		return nil
+	}
+
+	cfg := cfgs[0].(map[string]interface{})
+	enabled, _ := cfg["enabled"].(bool)
+	journeySettings := &platformclientv2.Journeyeventssettings{
+		Enabled: &enabled,
+	}
+
+	excludedQueryParams := lists.InterfaceListToStrings(cfg["excluded_query_parameters"].([]interface{}))
+	journeySettings.ExcludedQueryParameters = &excludedQueryParams
+
+	if keepUrlFragment, ok := cfg["should_keep_url_fragment"].(bool); ok && keepUrlFragment {
+		journeySettings.ShouldKeepUrlFragment = &keepUrlFragment
+	}
+
+	searchQueryParameters := lists.InterfaceListToStrings(cfg["search_query_parameters"].([]interface{}))
+	journeySettings.SearchQueryParameters = &searchQueryParameters
+
+	pageviewConfig := cfg["pageview_config"]
+	if value, ok := pageviewConfig.(string); ok {
+		if value != "" {
+			journeySettings.PageviewConfig = &value
+		}
+	}
+
+	if clickEvents := readSelectorEventTriggers(cfg["click_event"].([]interface{})); clickEvents != nil {
+		journeySettings.ClickEvents = clickEvents
+	}
+
+	if formsTrackEvents := readFormsTrackTriggers(cfg["form_track_event"].([]interface{})); formsTrackEvents != nil {
+		journeySettings.FormsTrackEvents = formsTrackEvents
+	}
+
+	if idleEvents := readIdleEventTriggers(cfg["idle_event"].([]interface{})); idleEvents != nil {
+		journeySettings.IdleEvents = idleEvents
+	}
+
+	if inViewportEvents := readSelectorEventTriggers(cfg["in_viewport_event"].([]interface{})); inViewportEvents != nil {
+		journeySettings.InViewportEvents = inViewportEvents
+	}
+
+	if scrollDepthEvents := readScrollPercentageEventTriggers(cfg["scroll_depth_event"].([]interface{})); scrollDepthEvents != nil {
+		journeySettings.ScrollDepthEvents = scrollDepthEvents
+	}
+
+	return journeySettings
+}
+
+func readSelectorEventTriggers(triggers []interface{}) *[]platformclientv2.Selectoreventtrigger {
+	if triggers == nil || len(triggers) < 1 {
+		return nil
+	}
+
+	results := make([]platformclientv2.Selectoreventtrigger, len(triggers))
+	for i, value := range triggers {
+		if trigger, ok := value.(map[string]interface{}); ok {
+			selector := trigger["selector"].(string)
+			eventName := trigger["event_name"].(string)
+			results[i] = platformclientv2.Selectoreventtrigger{
+				Selector:  &selector,
+				EventName: &eventName,
+			}
+		}
+	}
+
+	return &results
+}
+
+func readFormsTrackTriggers(triggers []interface{}) *[]platformclientv2.Formstracktrigger {
+	if triggers == nil || len(triggers) < 1 {
+		return nil
+	}
+
+	results := make([]platformclientv2.Formstracktrigger, len(triggers))
+	for i, value := range triggers {
+		if trigger, ok := value.(map[string]interface{}); ok {
+			selector := trigger["selector"].(string)
+			formName := trigger["form_name"].(string)
+			captureDataOnAbandon := trigger["capture_data_on_form_abandon"].(bool)
+			captureDataOnSubmit := trigger["capture_data_on_form_submit"].(bool)
+			results[i] = platformclientv2.Formstracktrigger{
+				Selector:                 &selector,
+				FormName:                 &formName,
+				CaptureDataOnFormAbandon: &captureDataOnAbandon,
+				CaptureDataOnFormSubmit:  &captureDataOnSubmit,
+			}
+		}
+	}
+
+	return &results
+}
+
+func readIdleEventTriggers(triggers []interface{}) *[]platformclientv2.Idleeventtrigger {
+	if triggers == nil || len(triggers) < 1 {
+		return nil
+	}
+
+	results := make([]platformclientv2.Idleeventtrigger, len(triggers))
+	for i, value := range triggers {
+		if trigger, ok := value.(map[string]interface{}); ok {
+			eventName := trigger["event_name"].(string)
+			idleAfterSeconds := trigger["idle_after_seconds"].(int)
+			results[i] = platformclientv2.Idleeventtrigger{
+				EventName:        &eventName,
+				IdleAfterSeconds: &idleAfterSeconds,
+			}
+		}
+	}
+
+	return &results
+}
+
+func readScrollPercentageEventTriggers(triggers []interface{}) *[]platformclientv2.Scrollpercentageeventtrigger {
+	if triggers == nil || len(triggers) < 1 {
+		return nil
+	}
+
+	results := make([]platformclientv2.Scrollpercentageeventtrigger, len(triggers))
+	for i, value := range triggers {
+		if trigger, ok := value.(map[string]interface{}); ok {
+			eventName := trigger["event_name"].(string)
+			percentage := trigger["percentage"].(int)
+			results[i] = platformclientv2.Scrollpercentageeventtrigger{
+				EventName:  &eventName,
+				Percentage: &percentage,
+			}
+		}
+	}
+
+	return &results
+}
+
+func readMessengerSettings(d *schema.ResourceData) *platformclientv2.Messengersettings {
+	value, ok := d.GetOk("messenger")
+	if !ok {
+		return nil
+	}
+
+	cfgs := value.([]interface{})
+	if len(cfgs) < 1 {
+		return nil
+	}
+
+	cfg := cfgs[0].(map[string]interface{})
+	enabled, _ := cfg["enabled"].(bool)
+	messengerSettings := &platformclientv2.Messengersettings{
+		Enabled: &enabled,
+	}
+
+	if styles, ok := cfg["styles"].([]interface{}); ok && len(styles) > 0 {
+		style := styles[0].(map[string]interface{})
+		if primaryColor, ok := style["primary_color"].(string); ok {
+			messengerSettings.Styles = &platformclientv2.Messengerstyles{
+				PrimaryColor: &primaryColor,
+			}
+		}
+	}
+
+	if launchers, ok := cfg["launcher_button"].([]interface{}); ok && len(launchers) > 0 {
+		launcher := launchers[0].(map[string]interface{})
+		if visibility, ok := launcher["visibility"].(string); ok {
+			messengerSettings.LauncherButton = &platformclientv2.Launcherbuttonsettings{
+				Visibility: &visibility,
+			}
+		}
+	}
+
+	if fileUploads, ok := cfg["file_upload"].([]interface{}); ok && len(fileUploads) > 0 {
+		fileUpload := fileUploads[0].(map[string]interface{})
+		if modesCfg, ok := fileUpload["mode"].([]interface{}); ok && len(modesCfg) > 0 {
+			modes := make([]platformclientv2.Fileuploadmode, len(modesCfg))
+			for i, modeCfg := range modesCfg {
+				if mode, ok := modeCfg.(map[string]interface{}); ok {
+					maxFileSize := mode["max_file_size_kb"].(int)
+					fileTypes := lists.InterfaceListToStrings(mode["file_types"].([]interface{}))
+					modes[i] = platformclientv2.Fileuploadmode{
+						FileTypes:     &fileTypes,
+						MaxFileSizeKB: &maxFileSize,
+					}
+				}
+			}
+
+			if len(modes) > 0 {
+				messengerSettings.FileUpload = &platformclientv2.Fileuploadsettings{
+					Modes: &modes,
+				}
+			}
+		}
+	}
+
+	return messengerSettings
+}
+
+func readCobrowseSettings(d *schema.ResourceData) *platformclientv2.Cobrowsesettings {
+	value, ok := d.GetOk("cobrowse")
+	if !ok {
+		return nil
+	}
+
+	cfgs := value.([]interface{})
+	if len(cfgs) < 1 {
+		return nil
+	}
+
+	cfg := cfgs[0].(map[string]interface{})
+
+	enabled, _ := cfg["enabled"].(bool)
+	allowAgentControl, _ := cfg["allow_agent_control"].(bool)
+	channels := lists.InterfaceListToStrings(cfg["channels"].([]interface{}))
+	maskSelectors := lists.InterfaceListToStrings(cfg["mask_selectors"].([]interface{}))
+	readonlySelectors := lists.InterfaceListToStrings(cfg["readonly_selectors"].([]interface{}))
+
+	return &platformclientv2.Cobrowsesettings{
+		Enabled:           &enabled,
+		AllowAgentControl: &allowAgentControl,
+		Channels:          &channels,
+		MaskSelectors:     &maskSelectors,
+		ReadonlySelectors: &readonlySelectors,
+	}
+}
